Refactor to split storage related into storage_operations, storage_tracking
